urlshortener/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
filelock==3.18.0
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
requests-file==2.1.0
sqlparse==0.5.3
tldextract==5.3.0
tzdata==2025.2
urllib3==2.5.0
